# Optimize binary quantization