@tobias-weaviate 2025-06-17
Thank you for the PR! Doing the processing in blocks was a great idea. I added a commit on top of the PR in an attempt to simplify things slightly for the sake of maintainability. Let me know if you have any comments. Otherwise I will try to get this merged ASAP.