@vojtechtoman 2025-06-17
@tobias-weaviate Note that the current version is slightly slower (about 20% on my machine) due to the relatively expensive if-check in the hot loop. That said, I agree it's better from maintenance perspective.