@tobias-weaviate 2025-06-17
I agree that it would be nice to avoid the if-check in an elegant way, but I also think that the patch is fast enough as it is. 

From my local benchmarks comparing the initial commit `cb3b73` and my commit `909937` I don't see any major differences in performance:

```
 goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/compressionhelpers
cpu: Apple M4 Pro
                                                  │ cb3b73.txt  │             909937.txt             │
                                                  │   sec/op    │   sec/op     vs base               │
BinaryQuantization/BinaryQuantization-dim-32-14     14.83n ± 2%   14.39n ± 0%  -2.93% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-64-14     24.22n ± 1%   25.02n ± 1%  +3.28% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-100-14    37.83n ± 1%   37.68n ± 1%       ~ (p=0.255 n=10)
BinaryQuantization/BinaryQuantization-dim-256-14    94.91n ± 2%   97.83n ± 3%  +3.08% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-500-14    174.6n ± 1%   177.5n ± 1%  +1.69% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-1024-14   353.0n ± 1%   359.2n ± 3%  +1.77% (p=0.007 n=10)
BinaryQuantization/BinaryQuantization-dim-4096-14   1.399µ ± 1%   1.399µ ± 2%       ~ (p=0.811 n=10)
geomean                                             101.5n        102.4n       +0.91%
```

But I do see a major benefit from your approach when comparing against the main branch (using the commit prior to the PR):

```
goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/compressionhelpers
cpu: Apple M4 Pro
                                                  │ b20d5fb.txt  │             cb3b73.txt              │
                                                  │    sec/op    │   sec/op     vs base                │
BinaryQuantization/BinaryQuantization-dim-32-14      20.74n ± 2%   14.83n ± 2%  -28.51% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-64-14      42.21n ± 2%   24.22n ± 1%  -42.62% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-100-14     71.38n ± 3%   37.83n ± 1%  -47.01% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-256-14    201.70n ± 0%   94.91n ± 2%  -52.95% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-500-14     395.2n ± 1%   174.6n ± 1%  -55.82% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-1024-14    830.4n ± 1%   353.0n ± 1%  -57.50% (p=0.000 n=10)
BinaryQuantization/BinaryQuantization-dim-4096-14    3.534µ ± 2%   1.399µ ± 1%  -60.43% (p=0.000 n=10)
geomean                        
```